gui: Create button "Get students data" and "Send"
button "Get students data" -> get the data of the excel file and show in
the textArea details about the students.

button "Send" -> send the email for each student. This button start
disabled, but when the button "Get students data" is press, "Send" will
be active and ready to send emails